+++ bsl-manager/src/main/java/com/bsl/service/plan/impl/SemiFinishedProdServiceImpl.java
@@ -623,6 +623,10 @@ public class SemiFinishedProdServiceImpl implements SemiFinishedProdService {
 		if(!DictItemOperation.产品状态_已入库.equals(oldBslProductInfo.getProdStatus())){
 			throw new BSLException(ErrorCodeInfo.错误类型_状态校验错误, "只有在库的产品才允许删除");
 		}
+		//删除需要备注原因
+		if(StringUtils.isBlank(oldBslProductInfo.getRemark())) {
+			throw new BSLException(ErrorCodeInfo.错误类型_状态校验错误, "请先修改产品备注，说明删除原因。");
+		}
 		//开始删除
 		int result = bslProductInfoMapper.deleteByPrimaryKey(prodId);
 		if(result<0){
@@ -639,6 +643,7 @@ public class SemiFinishedProdServiceImpl implements SemiFinishedProdService {
 		bslStockChangeDetailRaw.setProdType(DictItemOperation.产品类型_半成品);//产品类型
 		bslStockChangeDetailRaw.setRubbishWeight(oldBslProductInfo.getProdRelWeight());//重量
 		bslStockChangeDetailRaw.setInputuser(oldBslProductInfo.getProdInputuser());//录入人
+		bslStockChangeDetailRaw.setRemark(oldBslProductInfo.getRemark());//备注
 		bslStockChangeDetailRaw.setCrtDate(new Date());
 		int resultStockRaw = bslStockChangeDetailMapper.insert(bslStockChangeDetailRaw);
 		if(resultStockRaw<0){
