+++ bsl-manager/src/main/java/com/bsl/service/prodmanager/impl/ProdServiceImpl.java
@@ -889,6 +889,10 @@ public class ProdServiceImpl implements ProdService {
 		if(!DictItemOperation.产品状态_已入库.equals(oldBslProductInfo.getProdStatus())){
 			throw new BSLException(ErrorCodeInfo.错误类型_状态校验错误, "只有在库的产品才允许删除");
 		}
+		//删除需要备注原因
+		if(StringUtils.isBlank(oldBslProductInfo.getRemark())) {
+			throw new BSLException(ErrorCodeInfo.错误类型_状态校验错误, "请先修改产品备注，说明删除原因。");
+		}
 		//开始删除
 		int result = bslProductInfoMapper.deleteByPrimaryKey(prodId);
 		if(result<0){
@@ -906,6 +910,7 @@ public class ProdServiceImpl implements ProdService {
 		bslStockChangeDetailRaw.setProdType(DictItemOperation.产品类型_成品);//产品类型
 		bslStockChangeDetailRaw.setRubbishWeight(oldBslProductInfo.getProdRelWeight());//重量
 		bslStockChangeDetailRaw.setInputuser(oldBslProductInfo.getProdInputuser());//录入人
+		bslStockChangeDetailRaw.setRemark(oldBslProductInfo.getRemark());//备注
 		bslStockChangeDetailRaw.setCrtDate(new Date());
 		int resultStockRaw = bslStockChangeDetailMapper.insert(bslStockChangeDetailRaw);
 		if(resultStockRaw<0){
@@ -1356,6 +1361,10 @@ public class ProdServiceImpl implements ProdService {
 		if(!DictItemOperation.产品状态_已入库.equals(oldBslProductInfo.getProdStatus())){
 			throw new BSLException(ErrorCodeInfo.错误类型_状态校验错误, "只有在库的产品才允许删除");
 		}
+		//删除需要备注原因
+		if(StringUtils.isBlank(oldBslProductInfo.getRemark())) {
+			throw new BSLException(ErrorCodeInfo.错误类型_状态校验错误, "请先修改产品备注，说明删除原因。");
+		}
 		//开始删除
 		int result = bslProductInfoMapper.deleteByPrimaryKey(prodId);
 		if(result<0){
@@ -1373,6 +1382,7 @@ public class ProdServiceImpl implements ProdService {
 		bslStockChangeDetailRaw.setProdType(DictItemOperation.产品类型_成品);//产品类型
 		bslStockChangeDetailRaw.setRubbishWeight(oldBslProductInfo.getProdRelWeight());//重量
 		bslStockChangeDetailRaw.setInputuser(oldBslProductInfo.getProdInputuser());//录入人
+		bslStockChangeDetailRaw.setRemark(oldBslProductInfo.getRemark());//备注
 		bslStockChangeDetailRaw.setCrtDate(new Date());
 		int resultStockRaw = bslStockChangeDetailMapper.insert(bslStockChangeDetailRaw);
 		if(resultStockRaw<0){
